requirements/dev.txt
-r ./base.txt

black==22.10.0
Faker==15.0.0
factory-boy==3.2.1
pytest==7.1.3
pytest-django==4.5.2